wad: Change the types of some moby fields

+++ src/wad/level.h
@@ -211,11 +211,11 @@ struct MobyInstance {
 	s32 bolts;
 	s32 o_class;
 	f32 scale;
-	s16 draw_dist;
+	s32 draw_dist;
 	s32 update_dist;
 	Vec3f position;
 	Vec3f rotation;
-	s8 group;
+	s32 group;
 	s32 is_rooted;
 	f32 rooted_dist;
 	s32 occlusion;
